twister: move file scanning out of class
This is all static functions that can maintained outside of the class
and called only in one place, so move this to the testsuite file.
